refactor: more renaming and custom errors
