+++ store/smartAccount/errors.ts
@@ -1,6 +1,13 @@
 export class SmartAccountInfoNotFoundError extends Error {
   constructor() {
-    super('Smart account info is not persisted, thus unavailable');
+    super('Smart account info is not found');
     this.name = 'SmartAccountInfoNotFoundError';
+  }
+}
+
+export class SmartAccountInfoNotPersistedError extends Error {
+  constructor() {
+    super('Smart account info is not persisted');
+    this.name = 'SmartAccountInfoNotPersistedError';
   }
 }
